% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBeetle.R
\name{runBeetle}
\alias{runBeetle}
\title{Wrapper function to analyse the running beetle spread according to the hypothesis of bindings to wetness}
\usage{
runBeetle(cost = costraster,type=alt,occurence=carabLocs)
}
\arguments{
\item{iniparam}{the full set of params as provided initEnvironGIS(). please look into the INI file for further settings of the makePEak() mode and special settings.
peak.mode

(1) minmax: extracts local maxima altitudes from an arbitrary Digital
                    Elevation Model (DEM) (optionally you may filter the data)

(2) merged approach: extract peaks from a DEM by analyzing morphometry and
                    landscape forms using the algorithm of Wood et others. The peak landforms are
                    analyzed for MAximunm heights and this location is related
                    to external peak data from Harrys peaklist or OSM}

\item{myenv}{SAGA environment variables provided by initEnvironGIS()}

\item{fname.DEM}{name of georeferenced DEM data in GDAL format}
}
\value{
runBeetle basically returns a list of coordinates altitudes (and names)
that will be used to calculate the independence value.
}
\description{
Tibetan carabidae are spreading correlated to available humidity. Long term humidity is bound to surface  and ground water. It seems to be obvious that the estimation of spreeading speed and range is in no case a simple euclidian one. runBeetle provides a first better estimation using a cost or friction analysis assuming that the spread is following natural lines of wetness e.g. valleys, humidity gradients...
}
\examples{
#### Example to use runBeetle for a common analysis of the
    estimated spreading distances of an specifified area

# You need a georeferenced DEM (GDAL format) as data input.
# NOTE the existing projection of the data file has to be exactly the same
# as provided in target.proj4  variable in the ini file



}
\author{
Chris Reudenbach
}
\references{
Rasemann, S., (2004), Geomorphometrische Struktur eines mesoskaligen alpinen Geosystems, Asgard-Verlag, St. Augustin, Bonner Geographische Abhandlungen, Heft Nr. 111, URL: \url{http://hss.ulb.uni-bonn.de/2003/0211/0211.htm}

Wood, J.D., (1996), The geomorphological characterisation of digital elevation models. PhD Thesis, University of Leicester, UK

Schmidt, J, & A. Hewitt, (2004). Fuzzy land element classification from DTMs based on geometry and terrain position. Geoderma 121.3, S.243-256. URL: \url{http://home.shirazu.ac.ir/~kompani/geomorphology/geomorphology-lec-papers-mehr88/schmidt-fuzzylandsurfaceclassifi-geoderma2004.pdf}

Breitkreutz, H.: Gipfelliste. URL: \url{http://www.tourenwelt.info/commons/download/bergliste-komplett.kmz.php}

OSM natural: keys \url{http://wiki.openstreetmap.org/wiki/Key:natural}
}

