% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parseSpeciesList.R
\name{gettupel}
\alias{gettupel}
\title{parseSpeciesList is a prototype parser for running beetle species lists}
\usage{
gettupel(inputFile, short = TRUE)
}
\arguments{
\item{short}{logical parameter if TRUE (default) the function trys to get only the names and country codes. If FALSE the full text
  will put in the data frame.}

\item{inputTXT}{a Text of the specified format}
}
\description{
parseSpeciesList is a first prototype to parse running beetle
  species lists as provided by the beetle enthusiasts.

  This is a very raw and simple approach not really written in R style.
  Actually the parser looks line by line for keywords and special tags.
  Unfortunately it has to be performed line by line due to the fact that
  keywords for sepcies are missing and the rules are not always matching.
  So in a first try we use the "family", "genus" and "subgenus" as keywords. They are always placed in the beginning of a
  line. After "genus" or "subgenus" there is a single line for a single species.
  In the species textline we will find a more or less systematic list of country
  codes that indicate all countries with known occurrence of this special species.

  The resulting dataframe is a not normalized relation ( so it means a huge table with mostly redundant informations).

  It looks like:

  familiy;   genus;   subgenus;  species;     loctype; country\cr
  Carabidae; Carabus; Carabinae; irregularis; A:;      GE\cr
  Carabidae; Carabus; Carabinae; irregularis; N:;      CZ\cr
  .
  .
  .
}
\examples{
### examples parseSpeciesList ###

 ### we need the stringr lib
 library(stringr)

 ### first the basic parsing
 inputFile <- system.file("extdata", "species.chunk",   package="parseSpeciesList")
 df <- gettupel(inputFile)

 ###  now a basic mapping example  ####

 ###  we need some more libs ;)
   if (!require(downloader)) {install.packages("downloader")}
   if (!require(maptools)) {install.packages("maptools")}
   if (!require(sp)) {install.packages("sp")}
   if (!require(devtools)) {install.packages("devtools")}
   library(downloader)
   library(maptools)
   library(sp)
   library(devtools)
   if (!require(mapview)) {install_github("environmentalinformatics-marburg/mapview")}
   library(mapview)

 # first we need some spatial geometries for mapping the data
 download("http://thematicmapping.org/downloads/TM_WORLD_BORDERS-0.3.zip",dest="worldborders.zip", mode = "wb")

 # unzip it
   unzip ("worldborders.zip",exdir = "./")

 # read it to a spatialpointdataframe
   sPDF <- readShapePoly('TM_WORLD_BORDERS-0.3.shp')
 # join the world countries to our data (iso2 seems best but stil poor)
   sPDF2 <- joinData2Map(
   df
     , nameMap = sPDF
     , nameJoinIDMap = "ISO2"
     , nameJoinColumnData = "loc")

 # we have to project it
   proj4string(sPDF2) <- CRS("+init=epsg:4326")

 # and finally plot it with mapview
   mapView(sPDF2)
}
\author{
Chris Reudenbach
}

